mudança nas classes

+++ src/br/edu/ufersa/sei/model/VO/DiretorVO.java
@@ -3,14 +3,15 @@ package br.edu.ufersa.sei.model.VO;
 import br.edu.ufersa.sei.exception.InsertException;
 
 public class DiretorVO extends UsuarioVO{
-	private long idDir;
+	private int idDir;
 
 	//getters and setters
 	public long getIdDir() {
 		return idDir;
 	}
 
-	public void setIdDir(long idDir) throws InsertException {
+
+	public void setIdDir(int idDir) throws InsertException {
 		if(idDir >= 0) {
 			this.idDir = idDir;
 		}else {
